# docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "description": "This is the microservice for Sequspace achievements, handling shining stars and other awards.",
    "title": "Sequspace Achievements API",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "name": "API Support",
      "email": "support@sequspace.com"
    },
    "version": "1.0"
  },
  "servers": [
    {
      "url": "https://sequachievementapi.sequspace.com/api/v1",
      "description": "production server"
    },
    {
      "url": "http://localhost:{port}/api/v1",
      "variables": {
        "port": {
          "default": "4006"
        }
      }
    }
  ],
  "paths": {
    "/achievements/shining-star": {
      "get": {
        "description": "Get a list of all shining stars across the system",
        "produces": [
          "application/json"
        ],
        "tags": [
          "achievements"
        ],
        "summary": "Get all shining stars",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          }
        }
      },
      "post": {
        "description": "Create a new shining star achievement for a user with a certificate upload",
        "tags": [
          "achievements"
        ],
        "summary": "Create a new shining star achievement",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["user_id", "type", "description"],
                "properties": {
                  "user_id": {
                    "type": "integer",
                    "description": "User ID",
                    "example": 1
                  },
                  "type": {
                    "type": "string",
                    "description": "Achievement Type",
                    "enum": ["Task Master", "Consistency Star", "Speed Performer", "Deadline Champion", "Productivity Pro"]
                  },
                  "description": {
                    "type": "string",
                    "description": "Description of the achievement"
                  },
                  "certificate": {
                    "type": "string",
                    "format": "binary",
                    "description": "Certificate file to upload"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          }
        }
      }
    },
    "/achievements/shining-star/{user_id}": {
      "get": {
        "description": "Get a list of shining stars for a specific user",
        "produces": [
          "application/json"
        ],
        "tags": [
          "achievements"
        ],
        "summary": "Get user's shining stars",
        "parameters": [
          {
            "type": "integer",
            "description": "User ID",
            "name": "user_id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "apiKey",
        "name": "Authorization",
        "in": "header"
      }
    }
  }
}